8.May, stable

+++ project1/cards/views.py
@@ -2,17 +2,24 @@ import json
 import random
 
 from django.db.models import Case, When, Count, F, CharField
-from django.http import HttpResponseRedirect
+from django.http import HttpResponseRedirect, HttpResponse
 from django.shortcuts import get_object_or_404, redirect, render
 from django.urls import reverse_lazy, reverse
+from django.utils import timezone
 from django.views.generic import (
     ListView,
     CreateView,
-    UpdateView, DeleteView,
+    UpdateView,
+    DeleteView,
+    View,
 )
 
 from .forms import CardCheckForm, CardForm
-from .models import Cards, Mappings, Categories
+from .models import Cards, Mappings, Categories, ActiveStudy
+from django.core.cache import cache
+from cards.check_post_request import CheckReuqest, is_post_unique
+from django.db.models import Q
+
 
 
 class CardsListView(ListView):
@@ -22,20 +29,10 @@ class CardsListView(ListView):
     def get_queryset(self):
         slug = self.kwargs['slug']
         queryset = Cards.objects.filter(
-            mappings__category__slug=slug
-        ).annotate(
-            front_side=Case(
-                When(mappings__front_side=True, then=F('side1')),
-                default=F('side2'),
-                output_field=CharField(),
-            ),
-            back_side=Case(
-                When(mappings__front_side=True, then=F('side2')),
-                default=F('side1'),
-                output_field=CharField(),
-            ),
+            mappings__category__slug=slug,
+            mappings__is_back_side=False
         ).values(
-            'mappings__repetition', 'front_side', 'back_side', 'id', 'mappings', 'mappings__category__name'
+            'mappings__repetition', 'id', 'mappings', 'mappings__category__name', front_side=F('side1'), back_side=F('side2')
         ).order_by('mappings__repetition')
         return queryset
 
@@ -73,15 +70,21 @@ class CardCreateView(CreateView):
     success_url = reverse_lazy('card_create')
 
     def form_valid(self, form):
-        self.object = form.save(commit=False)
-        self.object.save()
-        Mappings.objects.create(card=self.object, category=form.cleaned_data['category'])
+        if is_post_unique(self.request):
+            self.object = form.save(commit=False)
+            self.object.save()
+            category = form.cleaned_data['category']
+            Mappings.objects.create(card=self.object, category=category)
+            if form.cleaned_data['is_two_sides']:
+                Mappings.objects.create(card=self.object, category=category, is_back_side=True)
         return HttpResponseRedirect(self.request.path)
 
     def get_form_kwargs(self):
         kwargs = super(CardCreateView, self).get_form_kwargs()
         kwargs['slug'] = self.kwargs.get('slug')
         return kwargs
+
+
 
 
 #
@@ -101,15 +104,17 @@ class CardCreateView(CreateView):
 class CardUpdateView(CardCreateView, UpdateView):
 
     def form_valid(self, form):
-        self.object = form.save(commit=False)
-        self.object.save()
         slug = self.kwargs['slug']
-        card = Mappings.objects.get(card=self.object).category = form.cleaned_data['category']
-        card.save()
+        if is_post_unique(self.request):
+            self.object = form.save(commit=False)
+            self.object.save()
+            card = Mappings.objects.get(card=self.object)
+            card.category = form.cleaned_data['category']
+            card.save()
         return HttpResponseRedirect(reverse('deck_content', args=[slug]))
 
 
-class CardDeleteView(DeleteView):
+class CardDeleteView(CheckReuqest, DeleteView):
     template_name = 'cards/delete_object.html'
     model = Cards
 
@@ -119,35 +124,63 @@ class CardDeleteView(DeleteView):
 
 
 
+def query_builder(*args, **kwargs):
+    conditions = Q(category__slug=kwargs['slug'])
+
+    if kwargs['study_mode'] == 'new':
+        conditions &= Q(study_mode='new') | Q(mem_rating__range=(1, 4), study_mode='new')
+
+    queryset = Mappings.objects.filter(conditions).order_by('mem_rating').limit(10).order_by('-upd_date')
 
 
-class BoxView(CardsListView):
+class BoxView(ListView):
     model = Cards
     template_name = 'cards/box.html'
     form_class = CardCheckForm
 
     def get_queryset(self):
-        queryset = super().get_queryset()
-        if 'rep' in self.kwargs and self.queryset:
-            rep = self.kwargs['rep']
-            queryset = queryset.filter(mappings__repetition=rep)
+        conditions = Q(mappings__category__slug=self.kwargs['slug'])
+        if self.request.GET.get('study_mode') == 'new':
+            conditions &= Q(mappings__review_date__isnull=True)
+        elif self.request.GET.get('study_mode') == 'review':
+            conditions &= Q(mappings__review_date__lt=timezone.now(), mappings__review_date__isnull=False)
+        # if 'rep' in self.kwargs:
+        #     conditions &= Q(mappings__repetition=self.kwargs['rep'])
+        self.kwargs['conditions'] = conditions
+        queryset = Cards.objects.filter(conditions)
+
+        queryset = queryset.annotate(
+            front_side=Case(
+                When(mappings__is_back_side=True, then=F('side2')),
+                default=F('side1'),
+                output_field=CharField(),
+            ),
+            back_side=Case(
+                When(mappings__is_back_side=True, then=F('side1')),
+                default=F('side2'),
+                output_field=CharField(),
+            ),
+        ).values(
+            'mappings__repetition', 'front_side', 'back_side', 'id', 'mappings', 'mappings__category__name'
+        ).order_by('mappings__repetition')
         return queryset
 
     def get_context_data(self, **kwargs):
         context = super().get_context_data(**kwargs)
         context['repetition'] = self.kwargs.setdefault('rep', 0)
         context['slug'] = self.kwargs['slug']
+        context['conditions'] = self.kwargs['conditions']
         if self.object_list:
             context['check_card'] = random.choice(self.object_list)
         return context
 
     def post(self, request, *args, **kwargs):
         form = self.form_class(request.POST)
-        if form.is_valid():
-            card = get_object_or_404(Mappings, card_id=form.cleaned_data['card_id'])
-            card.move(form.cleaned_data['rating'])
-
-        return redirect(request.META.get('HTTP_REFERER'))
+        if form.is_valid() and is_post_unique(request):
+            mappings = get_object_or_404(Mappings, pk=form.cleaned_data['mappings'])
+            mappings.move(form.cleaned_data['rating'])
+        success_url = self.request.get_full_path()
+        return HttpResponseRedirect(success_url)
 
 
 class DecksListView(ListView):
@@ -163,7 +196,7 @@ class DecksListView(ListView):
         return context
 
 
-class DeckCreateView(CreateView):
+class DeckCreateView(CheckReuqest, CreateView):
     model = Categories
     fields = ['name']
     template_name = 'cards/deck_form.html'
@@ -182,7 +215,7 @@ class DeckUpdateView(DeckCreateView, UpdateView):
 
 def deck_delete(request, pk):
     deck = get_object_or_404(Categories, pk=pk)
-    if request.method == 'POST':
+    if request.method == 'POST' and is_post_unique(request):
         cards = Cards.objects.filter(mappings__isnull=True)
         deck.delete()
         cards.delete()
@@ -190,9 +223,41 @@ def deck_delete(request, pk):
     return render(request,'cards/delete_object.html', {'deck_name': deck.name})
 
 
+def check_post_unique(view_func):
+    def wrapper(request, *args, **kwargs):
+        if request.POST:
+            token = request.POST.get('csrfmiddlewaretoken')
+            if cache.get(token):
+                print('doooouble')
+                request.POST = {}
+            cache.set(token, 'processed', timeout=50)
+        return view_func(request, *args, **kwargs)
+    return wrapper
+
+# @check_post_unique
 def test_contact(request):
     if request.method == 'POST':
         print(request.POST)
-        with open('log_requests.txt', "a") as f:
-            json.dump(request.POST, f, indent=4, ensure_ascii=False)
+        print(is_post_unique(request))
+        # if is_post_unique(request):
+        #     print('ok')
+        # else:
+        #     print('DOUBLE POST!!!!')
     return render(request, 'cards/test_contact.html', {})
+
+
+# class CheckReuqest(View):
+#     def post(self, request, *args, **kwargs):
+#         self.object = None
+#         token = request.POST.get('csrfmiddlewaretoken')
+#         if cache.get(token):
+#             return HttpResponseRedirect(self.request.path)
+#         cache.set(token, 'processed', timeout=50)
+#
+#         return super().post(request, *args, **kwargs)
+
+
+
+
+
+
